Move common_type() to dpnp_iface_manipulation.py
